adding voice cloning

+++ src/senselab/utils/functions.py
@@ -19,7 +19,7 @@ def _select_device_and_dtype(device_options: list[DeviceType] = [DeviceType.CPU,
         torch_dtype = torch.float16  # Using half precision for CUDA
     elif torch.backends.mps.is_available() and DeviceType.MPS in device_options:
         device = DeviceType.MPS
-        torch_dtype = torch.float16  # Using half precision for MPS if suitable
+        torch_dtype = torch.float32  # Default to float32 on MPS for better precision
     else:
         device = DeviceType.CPU
         torch_dtype = torch.float32  # Default to float32 on CPU for better precision
